Further speed up name validation
Bring the name checking regexes into LibXML and optimize them. There
is, for example, no use in quantifying `<.ident>`, and it's also
possible to organize them to match a bit faster in the common case.

While those bits could be upstreamed to the XML module, the other win is
just by avoiding the overhead of a nested regex, which is harder to get
without bringing the code into this module.

This gives a 20% improvement on the benchmark, which may be a big enough
win to justify this more invasive change.