Make `require` conditionally lock-protected
With merge of rakudo/rakudo#4917 and rakudo/rakudo#4934, lock-protecting
of `require` is not needed anymore. For the clarity of it, it never was
fully helpful as concurrent `require` calls in other modules or the
consuming application were still causing race conditions. But it was a
temporary relief and a solution when LibXML was the only user of
run-time module loading.