Fixing an intermediate state
More improvements to object construction:

- Added method `create` on `_Configurable` role to simplify creation of
  configurable objects where `:config` is required. It propagates
  current object's `$.config` to a newly created instance:
  `$node.create: LibXML::Element, ...;`
- Similarly, added a `box` method candidate with the same functionality:
  `$node.box: LibXML::Element, $raw, ...;`
- Construction methods are still somewhat of a mess, but generally they
  are being organized into layers. `new` is the lowest, mostly to be
  avoided by user; `create` is "low-level for user"; `box` is the
  high-level for item-classes.
- Similarly `LibXML::Config`'s `class-from` and `LibXML::Item`
  `box-class` are low- and high-level class resolution methods.

The construction subsystem is still considered WIP.

With regard to configuration, it has involuntarily became the center
object of an object (not class) tree. Normally, this role is given to
some kind of high-level aggregators like an application instance. But in
case of `libxml` the configuration is what gives shape to all objects in
a DOM. Even the document itself could be pulled out and reconstructed
with the help of `libxml2` API. Moreover, a sub-tree of DOM could be
sub-parsed using different configuration giving it slightly different
properties (embed XML into HTML, for example?). In the latter case one
would need to maintain reference to a different sub-root object which
might happen to be rather confusing. Hence, with all this in mind,
config-centric approach seems to be the right thing.

Aside of the above:

- Added `_Collection` role for the kind of objects which can contain
  children objects. Like, `LibXML::Node`, or `LibXML::HashMap`, etc. It
  is where currently `iterate-set` and `iterate-list` live in their
  method form.
- First introduction of config singleton. It's current purpose is to
  allow some tests not to use an explicitly created config object to
  create standalone elements. But it has a potential to become the
  replacement of module-based global config.