Don't use AttrX::Mooish
`libxml2` is rather nervous about OS thread changes in the middle of
some continuous operations. This happens when non-blocking `await` is
involved, and `AttrX::Moish` is using promises internaly to synchronize
some concurrent operations. For this reason it is about being better
safe than sorry.

Another side of the story is `LibXML::Config` where a user would rather
expect the globalish settings to be snapshotted as soon as a new
instance of the class is creted. But with lazy attributes this was not
the case:

    LibXML::Config.option = $val1;
    my $config = LibXML::Config.new;
    LibXML::Config.option = $val2;
    say $config.option; # $val2