Convert LibXML::Config to singleton-based approach
And make the singleton the default for all new configurable instances.

Also, the configuration class is now a `monitor` (`OO::Monitors`) what
makes it thread-safe with regard to the method calls. Note though that
assigning to a property is not thread-safe because it happens outside of
a method call, after property's container is returned to the calling
code. But this currently makes more sense as long as localized copies of
configs are used per-each parser (must be done manually!).