Add a `create` method candidate to be used with .=
It is distinguished by `:for`/`:from` named argument which must provide
a `LibXML::_Configurable` instance to pull a config from and use it for
the new object construction. I.e.:

    my LibXML::Element $e .= create: :for($doc);

is equivalent to

    mu LibXML::Element $e = $doc.create: LibXML::Element;

Though technically `:for` and `:from` are equivalent, they can be used
to distinguish the purpose of the newly created object: either it is
going to be added to the DOM where `$for` belongs to; or it is going to
be used as a standalone item but with the same configuration as `$from`.