Make LibXML::Item use :config for boxing (#97)
Move the call to $.raw.set-flags to a TWEAK submethod as it needs to be
called after $.raw.Reference has been called.
Any config supplied by upstream was previously ignored and class mapping
was expected to be done by `LibXML::Node`. This resulted in limited
abilities of user code in controlling how boxing is done. See #94.

Note that `LibXML::Node` still does its own class mapping because this
is the point of final decision and in many cases it is invoked directly.