Make LibXML::Item use :config for boxing (#97)
Any config supplied by upstream was previously ignored and class mapping
was expected to be done by `LibXML::Node`. This resulted in limited
abilities of user code in controlling how boxing is done. See #94.

Note that `LibXML::Node` still does its own class mapping because this
is the point of final decision and in many cases it is invoked directly.