Re-delegate Node's box method to Config and add caching support
A new config attribute `with-cache` defines if simple caching based on
node's `unique-key` should be used.

Method `box` is re-delegated to `LibXML::Config` to allow user code to
redefine boxing at the deep levels where control over `LibXML` internals
is normally impossible. The point is in allowing a user to fully
override the way boxing is done allowing for things like 3rd-party
caching, for example.

Also `class-from` now supports mapping from a raw type into a high-level
class. This is useful for cases when `$raw.type` is not availabe because
`$raw` is a typeobject, but we still need to know what class would be
the face of this native representation.