simplify variable argument process on generic-error handlers
For some reason the C wrapper code was marshalling the variable
arguments and pushing them forward for handling in Raku, instead
of just building the string with one of the v*printf functions,
specifically designed for the purpose.

Noticed while investigating #107, which is dieing in some of the
unnecessary Raku side argument processing.